+++ spinner/spinner_posix.go
@@ -1,4 +1,6 @@
-package ek
+//go:build linux || darwin
+
+package spinner
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 //                                                                                    //
@@ -7,5 +9,15 @@ package ek
 //                                                                                    //
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-// VERSION is current ek package version
-const VERSION = "14.1.9"
+import (
+	"github.com/essentialkaos/ek/v14/mathutil"
+	"github.com/essentialkaos/ek/v14/terminal/tty"
+)
+
+// ////////////////////////////////////////////////////////////////////////////////// //
+
+// getMaxDescSize returns the maximum character width available for the description text
+func getMaxDescSize() int {
+	w := tty.GetWidth()
+	return mathutil.B(w < 20, 9999, w-14)
+}