+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### [14.1.9](https://kaos.sh/ek/14.1.9)
+
+- **`[terminal]`** Fixed bug with reading data on Windows using `Read` and `ReadAnswer`
+- **`[spinner]`** Fixed Windows support
+
 ### [14.1.8](https://kaos.sh/ek/14.1.8)
 
 - **`[path]`** Added method `Sanitize`